Add some meta-tests
We need to have a test for exactly_equal(), to make sure it is correct.

We will need to figure out how to do meta-tests better in the future. I think
they should actually be separate from the main array_api_tests module. But I
can't figure out how to make it work, so for now, we just include it directly
and hard-code NumPy.