Add annotation stubs for annotation variables that weren't previously defined
If we ever end up using the annotations in the tests themselves we will likely
need to update the stub types in _types.py to be more sophisticated. For now,
though, the annotations are only included in the stubs for completeness and
convenience, and _types.py is only used to make pyflakes work.