Add ndindex() and promote_dtype() helper functions
The latter should probably be used in some more places in the dtype tests, but
for now I am not refactoring those. The conversion between dtype objects and
their string representation is finicky, and probably should be removed
entirely. It's not simple though, because the spec doesn't require dtype
objects to be usable as dict keys (and, e.g., NumPy does weird stuff like
float64 == None, which we have to be careful to guard against). Also we need
to make sure that the tests will work even if a dtype object is
_UndefinedStub.