Don't ignore exceptions in elementwise reference implementations
For some reason, "except OverflowError" was changed to "except Exception" in
e72184e5. For now I have removed the except entirely, but it's possible we may
need to keep the handling for OverflowError.

There are several issues with tests that this was masking, which I have not
fixed yet. Quite a few tests are not testing the complex implementations
correctly because they are using math instead of cmath, for example. See also
data-apis/array-api-compat#183.